requirements.txt
Flask==2.1.2
tensorflow-gpu==2.8.0
opencv-python==4.5.3.56
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
matplotlib==3.4.3
